Add logger and move RoleClaim class
