+++ 8.0/BlazorWebAssemblyStandaloneWithIdentity/BlazorWasmAuth/Identity/CookieAuthenticationStateProvider.cs
@@ -14,7 +14,8 @@ namespace BlazorWasmAuth.Identity
     /// Create a new instance of the auth provider.
     /// </remarks>
     /// <param name="httpClientFactory">Factory to retrieve auth client.</param>
-    public class CookieAuthenticationStateProvider(IHttpClientFactory httpClientFactory) : AuthenticationStateProvider, IAccountManagement
+    public class CookieAuthenticationStateProvider(IHttpClientFactory httpClientFactory, ILogger<CookieAuthenticationStateProvider> logger)
+        : AuthenticationStateProvider, IAccountManagement
     {
         /// <summary>
         /// Map the JavaScript-formatted properties to C#-formatted classes.
@@ -95,7 +96,10 @@ namespace BlazorWasmAuth.Identity
                     ErrorList = problemDetails == null ? defaultDetail : [.. errors]
                 };
             }
-            catch { }
+            catch (Exception ex)
+            {
+                logger.LogError(ex, "App error");
+            }
 
             // unknown error
             return new FormResult
@@ -133,7 +137,10 @@ namespace BlazorWasmAuth.Identity
                     return new FormResult { Succeeded = true };
                 }
             }
-            catch { }
+            catch (Exception ex)
+            {
+                logger.LogError(ex, "App error");
+            }
 
             // unknown error
             return new FormResult
@@ -214,7 +221,10 @@ namespace BlazorWasmAuth.Identity
                     authenticated = true;
                 }
             }
-            catch { }
+            catch (Exception ex)
+            {
+                logger.LogError(ex, "App error");
+            }
 
             // return the state
             return new AuthenticationState(user);
@@ -232,15 +242,6 @@ namespace BlazorWasmAuth.Identity
         {
             await GetAuthenticationStateAsync();
             return authenticated;
-        }
-
-        public class RoleClaim
-        {
-            public string? Issuer { get; set; }
-            public string? OriginalIssuer { get; set; }
-            public string? Type { get; set; }
-            public string? Value { get; set; }
-            public string? ValueType { get; set; }
         }
     }
 }
