+++ 8.0/BlazorWebAssemblyStandaloneWithIdentity/BlazorWasmAuth/Identity/Models/RoleClaim.cs
@@ -1,0 +1,11 @@
+namespace BlazorWasmAuth.Identity.Models
+{
+    public class RoleClaim
+    {
+        public string? Issuer { get; set; }
+        public string? OriginalIssuer { get; set; }
+        public string? Type { get; set; }
+        public string? Value { get; set; }
+        public string? ValueType { get; set; }
+    }
+}
